# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Mechanic Shop",
    "version": "1.0.0",
    "description": "API for managing mechanics, customers, service tickets and inventory."
  },
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
    }
  },
...
  "paths": {
    "/mechanics/admin/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Admin Login",
        "description": "Logs in admin and returns an authorization token.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/AdminLoginPayload" }
          }
        ],
        "responses": {
          "200": { "description": "Successful login", "schema": { "$ref": "#/definitions/LoginResponse" } },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/mechanics/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Mechanic Login",
        "description": "Logs in a mechanic and returns an authorization token.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/MechanicLoginPayload" }
          }
        ],
        "responses": {
          "200": { "description": "Successful login", "schema": { "$ref": "#/definitions/LoginResponse" } },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/mechanics/": {
      "post": {
        "tags": ["Mechanics"],
        "summary": "Add Mechanic",
        "description": "New mechanic created by admin.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/MechanicPayload" }
          }
        ],
        "responses": {
          "201": { "description": "Mechanic created", "schema": { "$ref": "#/definitions/MechanicResponse" } },
          "400": { "description": "Invalid request" }
        }
      },
      "get": {
        "tags": ["Mechanics"],
        "summary": "Get All Mechanics",
        "description": "Retrieves all mechanics with pagination.",
        "parameters": [
          { "name": "page", "in": "query", "type": "integer", "required": false },
          { "name": "per_page", "in": "query", "type": "integer", "required": false }
        ],
        "responses": {
          "200": {
            "description": "List of mechanics",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/MechanicResponse" } }
          }
        }
      }
    },
    "/mechanics/{id}": {
      "get": {
        "tags": ["Mechanics"],
        "summary": "Get Mechanic by ID",
        "description": "Looks up a specific mechanic and their information by ID.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": {
          "200": { "description": "Mechanic found", "schema": { "$ref": "#/definitions/MechanicResponse" } },
          "404": { "description": "Mechanic not found" }
        }
      },
      "put": {
        "tags": ["Mechanics"],
        "summary": "Update Mechanic",
        "description": "Update a mechanic information by ID.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "integer" },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/MechanicUpdatePayload" }
          }
        ],
        "responses": {
          "200": { "description": "Mechanic updated", "schema": { "$ref": "#/definitions/MechanicResponse" } },
          "400": { "description": "Invalid input" }
        }
      }
    },
    "/mechanics/admin/update/{id}": {
      "put": {
        "tags": ["Mechanics"],
        "summary": "Update Mechanic (Admin)",
        "description": "Admin updates a mechanics information by ID.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "integer" },
          { "in": "body", "name": "body", "schema": { "$ref": "#/definitions/MechanicPayload" } }
        ],
        "responses": {
          "200": { "description": "Mechanic updated", "schema": { "$ref": "#/definitions/MechanicResponse" } },
          "404": { "description": "Mechanic not found" }
        }
      }
    },
    "/mechanics/admin/delete/{id}": {
      "delete": {
        "tags": ["Mechanics"],
        "summary": "Delete Mechanic (Admin)",
        "description": "Admin deletes a mechanic by ID.",
        "security": [{ "Bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": {
          "204": { "description": "Mechanic deleted" },
          "404": { "description": "Mechanic not found" }
        }
      }
    },
    "/mechanics/profile": {
      "get": {
        "tags": ["Mechanics"],
        "summary": "Get Own Profile",
        "description": "Mechanic retrieves their own profile information.",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": { "description": "Mechanic profile", "schema": { "$ref": "#/definitions/MechanicResponse" } }
        }
      }
    },
    "/mechanics/my-tickets": {
      "get": {
        "tags": ["Mechanics"],
        "summary": "Get Assigned Tickets",
        "description": "Mechanic looks up all tickets assigned to themselves.",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": {
            "description": "List of tickets",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/TicketResponse" } }
          }
        }
      }
    },
    "/mechanics/dashboard": {
      "get": {
        "tags": ["Mechanics"],
        "summary": "Get Mechanic Dashboard",
        "description": "Provides stats and data for the mechanics dashboard.",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": { "description": "Dashboard data", "schema": { "type": "object" } }
        }
      }
    },
    "/mechanics/secure-data": {
      "get": {
        "tags": ["Mechanics"],
        "summary": "Get Secured Data",
        "description": "Mechanic retrieves own personal information.",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": { "description": "Secured data", "schema": { "type": "object" } }
        }
      }
    },
...
    "/service-tickets/{id}/mechanic": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Get Ticket by ID (Mechanic)",
        "description": "Mechanic pulls up any service ticket by ID.",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": {
            "description": "Ticket by ID",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/TicketResponse" } }
          }
        }
      }
    },
...
    "/inventory/mechanic": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Get All Inventory Parts (Mechanic)",
        "description": "Mechanic can see all parts in the inventory with pagination.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "page", "in": "query", "type": "integer", "required": false },
          { "name": "per_page", "in": "query", "type": "integer", "required": false }
        ],
        "responses": {
          "200": {
            "description": "List of inventory parts",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/InventoryPartResponse" } }
          }
        }
      }
    },
...
    "/inventory/mechanic/{id}": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Get an Inventory Part by ID (Mechanic)",
        "description": "Mechanic can search for a specific part in the inventory by ID.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { 
            "name": "id", 
            "in": "path",
            "required": true, 
            "type": "integer",
            "description": "Inventory Part ID"
            }
          ],
        "responses": {
          "200": {
            "description": "Inventory Part found",
            "schema": { "$ref": "#/definitions/InventoryPartResponse" }
          },
          "404": { "description": "Part not found" }
        }
      }
    },
...
  },
  "definitions": {
    "AdminLoginPayload": {
      "type": "object",
      "properties": {
        "username": { "type": "string", "example": "admin" },
        "password": { "type": "string", "example": "admin123" }
      },
      "required": ["username", "password"]
    },
    "MechanicLoginPayload": {
      "type": "object",
      "properties": {
        "email": { "type": "string", "example": "mechanic@example.com" },
        "password": { "type": "string", "example": "password123" }
      },
      "required": ["email", "password"]
    },
...
    "LoginResponse": {
      "type": "object",
      "properties": {
        "token": { "type": "string", "example": "jwt_token_here" }
      }
    },
    "MechanicPayload": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "email": { "type": "string" },
        "phone": { "type": "string", "example": "9874641234" },
        "address": { "type": "string" },
        "hours_worked": { "type": "integer", "example": 0 },
        "password": { "type": "string" }
      },
      "required": ["name", "email", "phone", "address", "password"]
    },
    "MechanicResponse": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "name": { "type": "string" },
        "email": { "type": "string" },
        "phone": { "type": "string" },
        "address": { "type": "string" },
        "hours_worked": { "type": "integer" }
      }
    },
    "MechanicUpdatePayload": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "email": { "type": "string" },
        "phone": { "type": "string", "example": "9874641234" },
        "address": { "type": "string" },
        "hours_worked": { "type": "integer", "example": 0 }
      }
    },
    "TicketResponse": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "description": { "type": "string" },
        "status": { "type": "string" },
        "customer_id": { "type": "integer" },
        "vehicle_id": { "type": "string" },
        "hours_worked": { "type": "integer" },
        "repair": { "type": "string" },
        "mechanics": {
          "type": "array",
          "items": { "$ref": "#/definitions/MechanicResponse" }
        }
      }
    },
...
    "InventoryPartResponse": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "name": { "type": "string" },
        "quantity": { "type": "integer" },
        "price": { "type": "number", "format": "float" }
      }
    },
...
  }
}